requirements.txt
numpy
pandas
scikit-learn
joblib
streamlit 
tabulapip 
imbalanced-learn
xgboost